% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geokz-package.R
\name{geokz-package}
\alias{geokz-package}
\title{Kazakhstani Administrative Divisions & Centers as The Geographic features}
\description{
Kazakhstani Administrative Divisions 0, 1 and 2 levels and Centers as The Geographic features
}
\note{
Languages available are:
\itemize{
\item \strong{"en"}: English
\item \strong{"kk"}: Kazakh
\item \strong{"ru"}: Russian
}

Although in Kazakhstan people mainly use codes \href{https://data.egov.kz/datasets/view?index=kato}{KATO}
(rus. - Classifier of Administrative Territorial Objects), an analogue of the Russian classifier \strong{OKATO},
we offer to apply:

\strong{ADM1_PCODE} = Two first characters of code according ISO 3166-2 ("KZ") +
\strong{AB} characters of \emph{KATO} (two first characters from \emph{TE}),  and

\strong{ADM2_PCODE} = Two first characters of code according ISO 3166-2 ("KZ") +
\strong{AB} + \strong{CD} + \strong{EF} characters of \emph{KATO} (two six characters from \emph{TE}),
for administrative units 1 level there are also geocodes according to \href{https://en.wikipedia.org/wiki/ISO_3166-2:KZ}{ISO 3166-2}.

\href{https://en.wikipedia.org/wiki/Almaty}{Almaty}, \href{https://en.wikipedia.org/wiki/Shymkent}{Shymkent} and
\href{https://en.wikipedia.org/wiki/Baykonur}{Baykonyr} has an specific status (City of Republican Significance),
but Baykonur (a city near the World's First Spaceport) considered as the city and the rayon
in a surrounding \href{https://en.wikipedia.org/wiki/Kyzylorda_Region}{Kyzylorda Region} on this \strong{POINT} dataset.
}
\section{Functions}{

\itemize{
\item \href{get_kaz_oblasts_map.html}{get_kaz_oblasts_map} - Get Oblasts & Cities of Republican Significance of Kazakhstan
as \code{sf} polygons.
\item \href{get_kaz_rayons_map.html}{get_kaz_rayons_map} - Get Rayons Oblasts & Cities of Oblast Significance of Kazakhstan
as \code{sf} polygons.
}
}

\section{Datasets with geographic features}{

\itemize{
\item \strong{kaz_adm?}: Geographic coverage (MULTIPOLYGON) of the country, regions (oblasts) and districts (rayons),
i.e. Administrative units 0, 1 and 2 levels respectively.
\item \strong{kaz_cnt1}: Geographic coverage (POINT) the Capital of Kazakhstan (Nur-Sultan),
Cities of Republican Significance (3)
and administrative centers(17)
of regions (oblasts).
\item \strong{natural_zones}: Data.frame with descriptions of Natural zones for districts (rayons),
i.e. Administrative units 2 levels respectively.
}

Several languages in datasets is available.
}

\keyword{"_PACKAGE"}
\keyword{internal}
