% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kaz_map.R
\name{get_kaz_rayons_map}
\alias{get_kaz_rayons_map}
\title{Get Rayons Oblasts & Cities of Oblast Significance of Kazakhstan as \code{sf} polygons}
\usage{
get_kaz_rayons_map(
                   KATO = NULL,        # Number
                   ADM1_EN = NULL,     # Character
                   ADM1_KK = NULL,     # Character
                   ADM1_RU = NULL,     # Character
                   ADM1_PCODE = NULL,  # Character
                   ADM2_EN = NULL,     # Character
                   ADM2_KK = NULL,     # Character
                   ADM2_RU = NULL,     # Character
                   ADM2_PCODE = NULL,  # Character
                   ISO_3166_2 = NULL,  # Character
                   crs = "+proj=lcc +lon_0=67 +lat_1=45 +lat_2=51 +ellps=krass",
                   Year = 2024L        # Integer
                   )
}
\arguments{
\item{KATO}{\strong{A vector} of Codes of \strong{Classifier of Administrative Territorial Objects} (rus.
\href{https://data.egov.kz/datasets/view?index=kato}{KATO}) for the required Oblasts &
Cities of Republican Significance of Kazakhstan or \code{NULL} to get all of them.}

\item{ADM1_EN}{\strong{A vector} of Names of Administrative units level 1 on English for the
required Oblasts & Cities of Republican Significance of Kazakhstan or \code{NULL} to get all of them.}

\item{ADM1_KK}{\strong{A vector} of Names of Administrative units level 1 on Kazakh (Cyrillic)
characters for the required Oblasts & Cities of Republican Significance of Kazakhstan or
\code{NULL} to get all of them.}

\item{ADM1_RU}{\strong{A vector} of Names of Administrative units level 1 on Russian (Cyrillic)
characters for the required Oblasts & Cities of Republican Significance of Kazakhstan or
\code{NULL} to get all of them.}

\item{ADM1_PCODE}{\strong{A vector} of \emph{modified} Codes of \strong{Classifier of Administrative Territorial Objects} (rus.
\href{https://data.egov.kz/datasets/view?index=kato}{KATO}) for the required Oblasts &
Cities of Republican Significance of Kazakhstan with \code{'KZ-'} prefix or
\code{NULL} to get all of them.}

\item{ADM2_EN}{\strong{A vector} of Names of Administrative units level 2 on English for the
required Rayons of Oblasts & Cities of Oblast Significance of Kazakhstan or \code{NULL} to get all of them.}

\item{ADM2_KK}{\strong{A vector} of Names of Administrative units level 2 on Kazakh (Cyrillic)
characters for the required Rayons of Oblasts & Cities of Oblast Significance of Kazakhstan or
\code{NULL} to get all of them.}

\item{ADM2_RU}{\strong{A vector} of Names of Administrative units level 2 on Russian (Cyrillic)
characters for the required Rayons of Oblasts & Cities of Oblast Significance of Kazakhstan
or \code{NULL} to get all of them.}

\item{ADM2_PCODE}{\strong{A vector} of \emph{modified} Codes of \strong{Classifier of Administrative Territorial Objects} (rus.
\href{https://data.egov.kz/datasets/view?index=kato}{KATO}) for the required Rayons of Oblasts &
Cities of Oblast Significance of Kazakhstan with \code{'KZ-'} prefix or
\code{NULL} to get all of them.}

\item{ISO_3166_2}{\strong{A vector} of Codes for the principal subdivisions coded in
\href{https://en.wikipedia.org/wiki/ISO_3166-2:KZ}{ISO 3166-2} or \code{NULL} to get all of them.}

\item{crs}{\strong{A value} of Coordinate Reference System as \href{https://en.wikipedia.org/wiki/EPSG_Geodetic_Parameter_Dataset}{EPSG (SRID)}
number (for example, 2502) or \code{proj4string} from the \href{https://proj.org/about.html}{PROJ.4} library
(for example, "+proj=longlat +datum=WGS84" as a projection specifier or
"+init=epsg:4326" as EPSG number) or \code{NULL} to get projection for A
\strong{Lambert Conformal Conic (LCC)} with \href{https://en.wikipedia.org/wiki/SK-42_reference_system}{Krasovsky 1940 ellipsoid}.  See \strong{Details}.}

\item{Year}{\strong{An integer} of Year for Administrative-Territorial divisions of the Country: 2018 or 2024 (default).}
}
\value{
A \code{sf} object with the requested geographic geometries.
}
\description{
Returns Geographical features of \href{https://en.wikipedia.org/wiki/Regions_of_Kazakhstan}{Rayons of Oblasts & Cities of Oblast Significance of Kazakhstan} -
Map of All Administrative units level 2 (the district units of principal units)
as polygons by specified codes or names or all of them in 2024 (default) or 2018 year.
}
\details{
When filter Geographical features you can use some names of Oblasts &
Cities of Republican Significance, Rayons of Oblasts & Cities of Oblast Significance
in English, Kazakh or Russian or National codes (\code{ADM1_PCODE} corresponding to level 1,
\code{ADM2_PCODE} or \href{https://data.egov.kz/datasets/view?index=kato}{KATO} corresponding to level 2)
or ISO International codes (\href{https://en.wikipedia.org/wiki/ISO_3166-2:KZ}{ISO 3166-2}
corresponding to level 1) (see \link{kaz_adm2_sf}).

Nur-Sultan, Almaty and Shymkent are considered as a set of regions on this dataset.

Baykonyr (a city near the World's First Spaceport) are not considered on this dataset.

Simultaneous use of several types of geographic features codes or
names in different languages \strong{is not allowed}.

For use non-latin characters in filters you need run \code{cat(stringi::stri_escape_unicode(x))} code R,
for example, \code{cat(stringi::stri_escape_unicode("\\u0428\\u044B\\u043C\\u043A\\u0435\\u043D\\u0442 (\\u049B\\u0430\\u043B\\u0430)"))}
that is "Шымкент (қала)" or
"\emph{Shymkent (city)}" in Kazakh language. See \code{stringi::\link[stringi]{stri_escape_unicode}} for details.

The \strong{Default Projection} uses \href{https://proj.org/operations/projections/lcc.html}{A Lambert Conformal Conic (LCC)}
with \emph{Krasovsky 1940 ellipsoid} and \emph{First standard parallel} - 45 N,
\emph{Second standard parallel} - 51 N, \emph{Longitude of projection center} - 67 E,
\emph{Latitude of projection center} - 48 N.
}
\examples{
kaz <- get_kaz_rayons_map()

plot(
  kaz["ADM1_EN"],
  main = "Rayons of Oblasts & Cities of Oblast Significance of Kazakhstan"
)


if (requireNamespace("dplyr", "tmap", quietly = TRUE)) {

library(dplyr)
library(tmap)

# Northern Regions

Northern_Region <-
  c(
    "Akmola",
    "Kostanay",
    "Pavlodar",
    "North Kazakhstan",
    "Nur-Sultan (city)"
  )

tmap::tm_shape(shp = get_kaz_oblasts_map(ADM1_EN = Northern_Region)) +
  tmap::tm_fill(
    "ADM1_EN",
    title = "Administrative \nunits level 1"
    ) +
tmap::tm_shape(
  shp = get_kaz_rayons_map(ADM1_EN = Northern_Region),
  is.master = TRUE
  ) +
  tmap::tm_polygons(
    "MAP_COLORS",
    palette = "Greys",
    alpha = 0.25
    ) +
tmap::tm_shape(get_kaz_oblasts_map()) +
  tmap::tm_borders(lwd = 2) +
  tmap::tm_text("ADM1_KK", size = 1.5, shadow = TRUE) +
  tmap::tm_format(
    format  = "World",
    title    = "Northern Region of Kazakhstan",
    bg.color = "white"
    )

# Simultaneous use of several types of International & National codes or
# names in different languages **is allowed** with binding geographic features

# Torgay Oblast is disbanded Region of the Kazakh Soviet Socialist Republic from 1970 to 1988
# and of independent Kazakhstan and from 1990 to 1997 <https://en.wikipedia.org/wiki/Torgay_Region>.

Torgay_Region <-
  get_kaz_rayons_map(ADM2_EN = c(
    "Esil",
    "Zhaksy",
    "Zharkain")) \%>\%
  dplyr::filter(ADM1_EN == "Akmola") \%>\%  # There are three rayons named Esil by Esil (Ishim) river
  dplyr::bind_rows(
    get_kaz_rayons_map(ADM2_PCODE = c(
      "KZ391600",                         # Arkalyk city was a Center of Torgay Oblast
      "KZ393400",                         # Amangeldi rayon
      "KZ394200")                         # Zhangeldi rayons
      )
  )

center_sf <-
  data.frame(
    NAME = c("Arkalyk"),
    BRANCH = c("The Center"),
    LINK = c("https://en.wikipedia.org/wiki/Arkalyk"),
    LAT = c(50.24861),
    LON = c(66.91139)
  ) \%>\%
  sf::st_as_sf(
    coords = c("LON", "LAT"), # columns with geometry by x - "LON" and y - "LAT"
    crs = "+proj=longlat +datum=WGS84"
  )

to_bb <-
  Torgay_Region \%>\%
  sf::st_bbox() \%>\%
  sf::st_as_sfc() \%>\%
  sf::st_buffer(dist = 50000)

tmap::tm_shape(Torgay_Region) +
  tmap::tm_polygons(
    col = "ADM2_EN",
    border.col = "grey",
    title = "Administrative \nunits level 2",
    labels = Torgay_Region$ADM2_KK
  ) +
  tmap::tm_text("ADM2_KK", col = "gray10", size = 0.9, shadow = TRUE) +
tmap::tm_shape(
  shp  = get_kaz_oblasts_map(),
  bbox = to_bb
  ) +
  tmap::tm_borders(
    col = "red",
    lwd = 4
  ) +
  tmap::tm_text("ADM1_KK", col = "gray10", size = 1.2, shadow = TRUE) +
tmap::tm_shape(center_sf) +
  tmap::tm_dots(size = 1) +
tmap::tm_graticules(lines = FALSE) +
  tmap::tm_format(format = 'World') +
  tmap::tm_layout(
    main.title = "Disbanded Torgay Oblast (1970-1988, 1990-1997)",
    main.title.size = 0.9,
    legend.position = c("left", "top")
  ) +
  tmap::tm_credits('Source: package {geokz}', position = c('right', 'bottom'), size = 1.2)

}    # The end for {dplyr} & {tmap} package

\dontrun{
  get_kaz_rayons_map("a")
  get_kaz_rayons_map(crs = "a")
}

}
\seealso{
You can upload dataset \link{kaz_adm2_sf} for Rayons of Oblasts & Cities of Oblast Significance
of Kazakhstan directly, but you must manually filter the Geographic Features
you require and set the suitable \href{https://en.wikipedia.org/wiki/List_of_map_projections}{Cartographic Projection}.

For maps of Rayons of Oblast & City of Oblast Significance in 2024 Version please use function
\link{get_kaz_rayons_map}\verb{()} or dataset \link{kaz_adm2_sf}.

For maps of Rayons of Oblast & City of Oblast Significance in 2018 Version please use function
\link{get_kaz_rayons_map}\code{(Year = 2018L)} or dataset \link{kaz_adm2_2018_sf}.

\href{../doc/making_maps.html}{package vignette} or \code{vignette("making_maps", package = "geokz")}

Other general polygon functions: 
\code{\link{get_kaz_oblasts_map}()}
}
\author{
Alexander Rodionov \email{a.rodionoff@gmail.com}, \strong{ORCID} = "0000-0003-2028-5421".
}
\concept{general polygon functions}
