% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{natural_zones}
\alias{natural_zones}
\title{Administrative units level 2 of Kazakhstan by Zones according to the natural conditions in 2020}
\format{
A data frame object with 216 rows and fields:
\itemize{
\item KATO: Classifier of Administrative Territorial Objects (rus.
\href{https://data.egov.kz/datasets/view?index=kato}{KATO})
\item ZONE_EN: Name of Zone on English (Latin) characters
\item ZONE_KK: Name of Zone on Kazakh (Cyrillic) characters
\item ZONE_RU: Name of Zone on Russian (Cyrillic) characters
\item ADM1_RU: Name of Administrative units level 1 on Russian (Cyrillic) characters
\item ADM2_RU: Name of Administrative units level 2 on Russian (Cyrillic) characters
\item ADM2_PCODE: Two first characters of code according ISO 3166-2 ("KZ") + (\strong{AB} characters of \emph{KATO} + \strong{CD} characters of \emph{KATO} + \strong{EF} characters of \emph{KATO}) or six first characters from \emph{TE}
}
}
\source{
\href{https://adilet.zan.kz/rus/docs/V2000021691#z70}{The list of administrative-territorial regions according to the natural conditions of the Republic of Kazakhstan}.
}
\usage{
data(natural_zones)
}
\description{
A \code{data.frame} object including the list of Administrative units level 2 of Kazakhstan (2018 Version)
by Zones according to the natural conditions.
}
\examples{
data("natural_zones")

# Use {tmap} package for draw of Thematic Maps

if (requireNamespace("tmap", "dplyr", "knitr", quietly = TRUE)) {

library(tmap)
library(dplyr)
library(knitr)

natural_zones \%>\%
  dplyr::count(ZONE_EN) \%>\%
  dplyr::bind_rows(., data.frame( ZONE_EN = "TOTAL", n = nrow(natural_zones))) \%>\%
  knitr::kable(.,
    caption = "Administrative units level 2 by Zones according to <b>natural conditions</b>",
    format.args = list(big.mark = ' ')
  )

dplyr::inner_join( x = get_kaz_rayons_map(),
                   y = natural_zones[, c("ADM2_PCODE", "ZONE_EN")],
                   by = c("ADM2_PCODE" = "ADM2_PCODE") ) \%>\%
  tmap::qtm(
    shp = .,
    fill = "ZONE_EN",
    fill.title = "Zones according \nto natural conditions",
    main.title = "Zones are distinguished according to natural conditions",
    format = "World"
  )

}  # The end for {tmap} package


# Use {ggplot2} package for draw of Advanced Maps as `sf` objects

if (requireNamespace("ggplot2", "dplyr", quietly = TRUE)) {

library(ggplot2)
library(dplyr)

Zones_EN_labels <-
  c("Steppe", "Dry-steppe", "Foothill-desert-steppe", "Desert",
    "South Siberian mountain and foothill", "Semi-desert")

Zones_pallete <- c("yellowgreen", "khaki", "darkolivegreen1", "peachpuff2",
                   "olivedrab", "navajowhite") \%>\%
  stats::setNames(Zones_EN_labels)

natural_zones_df <-
system.file("extdata", "kaz_zones.csv", package = "geokz", mustWork = TRUE) \%>\%
utils::read.csv(file = ., encoding = "UTF-8")


dplyr::inner_join( x = get_kaz_rayons_map(),
                 y = natural_zones_df[, c("ADM2_PCODE", "ZONE_EN")],
                 by = c("ADM2_PCODE" = "ADM2_PCODE") ) \%>\%
ggplot2::ggplot(data = .) +
  ggplot2::geom_sf(mapping = ggplot2::aes(fill = ZONE_EN)) +  # Rayons - Zones by natural conditions
  ggplot2::geom_sf(                                           # Boundaries of Kazakhstani Oblasts
    data  = sf::st_geometry(get_kaz_oblasts_map()),
    fill = NA,
    color = "red",
    size  = 1.25,
    show.legend = FALSE
  ) +
  ggplot2::scale_fill_manual(
    values = Zones_pallete,
    name = "Zones according \nto natural conditions"
  ) +
  ggplot2::theme_light() +
  ggplot2::labs(title = "Zones are distinguished according to natural conditions",
                x = NULL,
                y = NULL,
                caption = "Source: package {geokz}"
  )

}  # The end for {ggplot2} package

}
\seealso{
\link{get_kaz_rayons_map}

\href{/doc/making_maps.html#natural_conditions}{package vignette} or \code{vignette("making_maps", package = "geokz")}
}
\concept{datasets with Natural Zones}
\keyword{datasets}
