% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{natural_zones}
\alias{natural_zones}
\title{Administrative units level 2 of Kazakhstan by Zones according to the natural conditions}
\format{
A data frame object with 216 rows and fields:
\itemize{
\item KATO: Classifier of Administrative Territorial Objects (rus.
\href{https://data.egov.kz/datasets/view?index=kato}{KATO})
\item ZONE_EN: Name of Zone on English (Latin) characters
\item ZONE_KK: Name of Zone on Kazakh (Cyrillic) characters
\item ZONE_RU: Name of Zone on Russian (Cyrillic) characters
\item ADM1_RU: Name of Administrative units level 1 on Russian (Cyrillic) characters
\item ADM2_RU: Name of Administrative units level 2 on Russian (Cyrillic) characters
\item ADM2_PCODE: Two first characters of code according ISO 3166-2 ("KZ") + (\strong{AB} characters of \emph{KATO} + \strong{CD} characters of \emph{KATO} + \strong{EF} characters of \emph{KATO}) or six first characters from \emph{TE}
}
}
\source{
\href{https://adilet.zan.kz/rus/docs/V2000021691#z70}{The list of administrative-territorial regions according to the natural conditions of the Republic of Kazakhstan}.
}
\usage{
data(natural_zones)
}
\description{
A \code{data.frame} object including the list of Administrative units level 2 of Kazakhstan
by Zones according to the natural conditions.
}
\examples{
data("natural_zones")

# Use {tmap} package for draw of Thematic Maps

if (requireNamespace("tmap", "dplyr", "knitr", quietly = TRUE)) {

library(tmap)
library(dplyr)
library(knitr)

natural_zones \%>\%
  dplyr::count(ZONE_EN) \%>\%
  dplyr::bind_rows(., data.frame( ZONE_EN = "TOTAL", n = nrow(natural_zones))) \%>\%
  knitr::kable(.,
    caption = "Administrative units level 2 by Zones according to <b>natural conditions</b>",
    format.args = list(big.mark = ' ')
  )

dplyr::inner_join( x = get_kaz_rayons_map(),
                   y = natural_zones[, c("ADM2_PCODE", "ZONE_EN")],
                   by = c("ADM2_PCODE" = "ADM2_PCODE") ) \%>\%
  tmap::qtm(
    shp = .,
    fill = "ZONE_EN",
    fill.title = "Zones according \nto natural conditions",
    main.title = "Zones are distinguished according to natural conditions of Kazakhstan",
    format = "World"
  )

}  # The end for {tmap} package
}
\seealso{
\link{get_kaz_rayons_map}

\href{/doc/making_maps.html#natural_conditions}{package vignette} or \code{vignette("making_maps", package = "geokz")}
}
\concept{datasets with Natural Zones}
\keyword{datasets}
